scheduler: rework from thread to subroutine since instcache is often invalid
The cluster controller (presumably for scability reasons) makes little attempt
to keep the instance cache properly updated.  See how it handles
terminateInstance for an example--if instance isn't in cache, tell all nodes
to terminate the specificed instance.

Anyway this is clearly problematic for a scheduler that needs accurate full
knowledge of at least the instances that are running!

So instead of a separate thread that periodically runs the scheduler,
we hijack the existing monitor_thread and immediately after refreshing the
resource and instance caches, run the scheduler.

Messes up the periodic nature of the scheduler somewhat, but working
scheduler (actually has valid instance information!) is more important.