node: (HACK) issue kill if can't find instance to terminate
This piggybacks on previous commit that has cluster broadcast
terminate requests, so that if libvirt claims we don't have it
(for whatever reason), run a silly piped command to attempt to
kill any remaining kvm processes for that instance-id.

Idea is that when user says 'terminate', that vm is gone for sure.

NOTE: I have yet to see this actually help, as circumstances that
motivated its creation I have yet to encounter.. so ymmv, etc.